+++ index.js
@@ -6,14 +6,14 @@ export { LiistColors } from './src/atoms/LiistColors.js';
 export { LiistBttn } from './src/atoms/LiistBttn.js';
 export { LiistBttnOld } from './src/atoms/LiistBttnOld.js';
 export { LiistDaadBttn } from './src/atoms/LiistDaadBttn.js';
-export { LiistSection } from './src/atoms/LiistSection.js';
 /*
- * MOLECULES
- */
+* MOLECULES
+*/
 export { LiistAuthor } from './src/molecules/LiistAuthor.js';
 export { LiistCheckItem } from './src/molecules/LiistCheckItem.js';
 export { LiistPlaceCard } from './src/molecules/LiistPlaceCard.js';
 export { LiistPlaceCardOld } from './src/molecules/LiistPlaceCardOld.js';
+export { LiistSection } from './src/molecules/LiistSection.js';
 export { LiistToggleBttn } from './src/molecules/LiistToggleBttn.js';
 /*
  * ORGAMISMS
