+++ stories/2_molecules/LiistSection.stories.mdx
@@ -1,8 +1,8 @@
 import { Story, Preview, Meta, Props, html } from '@open-wc/demoing-storybook';
-import '../../src/atoms/LiistSection.js';
+import '../../src/molecules/LiistSection.js';
 
 <Meta
-  title="atoms|LiistSection"
+  title="molecules|LiistSection"
   parameters={{
     component: "liist-section",
     options: { selectedPanel: "storybookjs/knobs/panel" }
